+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
   "lib/vector-wasm",
   "lib/k8s-test-framework",
   "lib/prometheus-parser",
+  "lib/api_client"
 ]
 
 [dependencies]
@@ -47,6 +48,7 @@ codec = { path = "lib/codec" }
 file-source = { path = "lib/file-source", optional = true }
 tracing-limit = { path = "lib/tracing-limit" }
 prometheus-parser = { path = "lib/prometheus-parser", optional = true }
+api_client = { path = "lib/api_client", optional = true }
 
 # Tokio / Futures
 futures01 = { package = "futures", version = "0.1.25" }
@@ -98,12 +100,6 @@ smpl_jwt = { version = "0.5.0", optional = true }
 # API
 async-graphql = { version = "1.18.2", optional = true }
 async-graphql-warp = { version = "1.18.1", optional = true }
-
-# API client
-graphql_client = { version = "0.9.0", optional = true }
-reqwest = { version = "0.10.6", features = ["json"], optional = true }
-tokio-tungstenite = {version = "0.11.0", optional = true }
-weak-table = { version = "0.3.0", optional = true }
 
 # External libs
 derivative = "1.0"
@@ -230,6 +226,7 @@ default-cmake = ["api", "api_client", "sources", "transforms", "sinks", "vendor-
 # Default features for *-pc-windows-msvc
 # TODO: Enable SASL https://github.com/timberio/vector/pull/3081#issuecomment-659298042
 default-msvc = ["api", "api_client", "sources", "transforms", "sinks", "vendor-openssl", "vendor-libz", "leveldb", "rdkafka-cmake"]
+no-api-client = ["api", "sources", "transforms", "sinks", "vendor-openssl", "vendor-libz", "leveldb", "rdkafka-cmake"]
 
 # Target specific release features.
 # The `make` tasks will select this according to the appropriate triple.
@@ -264,14 +261,6 @@ kubernetes = ["k8s-openapi", "evmap"]
 api = [
   "async-graphql",
   "async-graphql-warp",
-]
-
-# API client
-api_client = [
-  "graphql_client",
-  "reqwest",
-  "tokio-tungstenite",
-  "weak-table",
 ]
 
 # Sources
