+++ lib/api_client/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "api_client"
+version = "0.1.0"
+authors = ["Vector Contributors <vector@timber.io>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+
+# Serde
+serde = { version = "1.0.116", features = ["derive"] }
+serde_json = { version = "1.0.33", features = ["raw_value"] }
+
+# Tokio / Futures
+futures = { version = "0.3", default-features = false, features = ["compat", "io-compat"] }
+tokio = { version = "0.2.13", features = ["blocking", "fs", "signal", "io-std", "macros", "rt-core", "rt-threaded", "uds", "sync"] }
+
+# GraphQL
+graphql_client = "0.9.0"
+
+# HTTP / WebSockets
+reqwest = "0.10.6"
+tokio-tungstenite = "0.11.0"
+
+# External libs
+weak-table = "0.3.0"
+url = "2.1.1"
+uuid = { version = "0.8", features = ["serde", "v4"] }
+chrono = { version = "0.4.6", features = ["serde"] }
+
+[dev-dependencies]
+vector = { version = "0.11.0", path = "../..", default-features = false, features = ["no-api-client"] }
