+++ lib/api_client/graphql/queries/health.graphql
