+++ lib/api_client/graphql/subscriptions/events_processed_metrics.graphql
