+++ lib/api_client/graphql/subscriptions/heartbeat.graphql
