+++ lib/api_client/graphql/subscriptions/uptime_metrics.graphql
