+++ lib/api_client/src/client/mod.rs
@@ -1,0 +1,5 @@
+pub mod http;
+pub mod ws;
+
+pub use http::*;
+pub use ws::*;
