+++ lib/api_client/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod client;
+pub mod query;
+pub mod subscription;
