+++ lib/api_client/src/query.rs
@@ -1,0 +1,9 @@
+use graphql_client::GraphQLQuery;
+
+#[derive(GraphQLQuery)]
+#[graphql(
+    schema_path = "graphql/schema.json",
+    query_path = "graphql/queries/health.graphql",
+    response_derives = "Debug"
+)]
+pub struct HealthQuery;
