--- src/api_client/mod.rs
@@ -1,5 +1,0 @@
-mod client;
-mod subscription;
-
-pub use client::*;
-pub use subscription::*;
